requirements.txt
Flask==3.0.0
stripe==7.10.0
gspread==6.0.2
oauth2client==4.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
